Moved measurement functions to separate module
